[MISRA][FRAMEWORK][MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012] -Fixed misra warning by typecasting the expression using unsigned long

+++ source/framework/vx_image.c
@@ -1747,7 +1747,7 @@ VX_API_ENTRY vx_status VX_API_CALL vxCopyImagePatch(
 
         if (image_addr->stride_x == 0)
         {
-            pImageLine = &(pImagePtr[((start_y*(vx_uint32)image_addr->stride_y)/image_addr->step_y) + (((start_x*12UL)/8UL)/image_addr->step_x)]);
+            pImageLine = &(pImagePtr[(((unsigned long)start_y*(vx_uint32)image_addr->stride_y)/image_addr->step_y) + (((start_x*12UL)/8UL)/image_addr->step_x)]);
 
             if (user_addr->stride_x == 1)
             {
